+++ tests/integration/article.test.js
@@ -31,138 +31,23 @@ before(done => {
     });
 });
 
-describe("# Articles endpoints", () => {
-  let createdArticle;
-  describe("Create a new article", () => {
-    it("should create the article and return the success message", done => {
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .set("Authorization", `Bearer ${token}`)
-        .send(newArticle)
-        .end((err, res) => {
-          createdArticle = res.body;
-          validSlug = res.body.article.slug;
-          expect(res.status).equals(CREATED);
-          expect(res.body.message).to.contain("Article created");
-          expect(res.body).to.haveOwnProperty("article");
-          expect(res.body.article).to.haveOwnProperty("createdAt");
-          expect(res.body.article).to.haveOwnProperty("title");
-          expect(res.body.article.title).to.contain(newArticle.title);
-          expect(res.body.article.readTime).equals(1);
-          done();
-        });
-    });
-
-    it("should deny the request if no access-token provided ", done => {
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .send(newArticle)
-        .end((err, res) => {
-          expect(res.status).equals(UNAUTHORIZED);
-          expect(res.body.message).to.contain("Please provide a token");
-          done();
-        });
-    });
-
-    it("should decline creating the article if no title provided", done => {
-      const { title, ...rest } = newArticle;
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .set("Authorization", `Bearer ${token}`)
-        .send(rest)
-        .end((err, res) => {
-          expect(res.status).equals(BAD_REQUEST);
-          expect(res.body.message).to.contain('"title" is required');
-          done();
-        });
-    });
-
-    it("should decline creating the article if no body provided", done => {
-      const { body, ...rest } = newArticle;
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .set("Authorization", `Bearer ${token}`)
-        .send(rest)
-        .end((err, res) => {
-          expect(res.status).equals(BAD_REQUEST);
-          expect(res.body.message).to.contain('"body" is required');
-          done();
-        });
-    });
-
-    it("should decline creating the article if no description provided", done => {
-      const { description, ...rest } = newArticle;
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .set("Authorization", `Bearer ${token}`)
-        .send(rest)
-        .end((err, res) => {
-          expect(res.status).equals(BAD_REQUEST);
-          expect(res.body.message).to.contain('"description" is required');
-          done();
-        });
-    });
-
-    it("should create an article with no tagsList", done => {
-      const { tagsList, ...rest } = newArticle;
-      chai
-        .request(app)
-        .post("/api/v1/articles")
-        .set("Authorization", `Bearer ${token}`)
-        .send(rest)
-        .end((err, res) => {
-          expect(res.status).equals(CREATED);
-          expect(res.body.message).to.contain("Article created");
-          expect(res.body).to.haveOwnProperty("article");
-          expect(res.body.article).to.haveOwnProperty("createdAt");
-          expect(res.body.article).to.haveOwnProperty("title");
-          expect(res.body.article.title).to.contain(newArticle.title);
-          done();
-        });
-    });
-  });
-  describe("Report an article endpoint", () => {
-    it("should be report an article", done => {
-      chai
-        .request(app)
-        .put(`/api/v1/articles/${validSlug}/report`)
-        .set("Authorization", `Bearer ${token}`)
-        .send({ description: "abusive" })
-        .end((err, res) => {
-          expect(res.status).equals(OK);
-          expect(res.body.message).to.contain("Article reported");
-          done();
-        });
-    });
-    it("should report an article", done => {
-      chai
-        .request(app)
-        .put(`/api/v1/articles/${validSlug}/report`)
-        .set("Authorization", `Bearer ${token}`)
-        .send({ description: "abusive" })
-        .end((err, res) => {
-          expect(res.status).equals(CONFLICT);
-          expect(res.body.message).to.contain("Article already reported");
-          done();
-        });
-    });
-    it("should should ask for description", done => {
-      chai
-        .request(app)
-        .put(`/api/v1/articles/${validSlug}/report`)
-        .set("Authorization", `Bearer ${token}`)
-        .send({ description: "" })
-        .end((err, res) => {
-          expect(res.status).equals(BAD_REQUEST);
-          expect(res.body.message).to.contain("Please, give a reason");
-          done();
-        });
-    });
+describe("Create a new article", () => {
+  it("should create the article and return the success message", done => {
+    chai
+      .request(app)
+      .post("/api/v1/articles")
+      .set("Authorization", `Bearer ${token}`)
+      .send(newArticle)
+      .end((err, res) => {
+        expect(res.status).equals(CREATED);
+        expect(res.body.message).to.contain("Article created");
+        expect(res.body).to.haveOwnProperty("article");
+        expect(res.body.article).to.haveOwnProperty("createdAt");
+        expect(res.body.article).to.haveOwnProperty("title");
+        expect(res.body.article.title).to.contain(newArticle.title);
+        expect(res.body.article.readTime).equals(1);
+        done();
+      });
   });
 
   it("should deny the request if no access-token provided ", done => {
@@ -185,9 +70,8 @@ describe("# Articles endpoints", () => {
       .set("Authorization", `Bearer ${token}`)
       .send(rest)
       .end((err, res) => {
-        expect(res.status).equals(OK);
-        expect(res.body.message).to.contain("Article ready to be posted on twitter");
-        expect(res.body.response.spawnfile).to.equal("open");
+        expect(res.status).equals(BAD_REQUEST);
+        expect(res.body.message).to.contain('"title" is required');
         done();
       });
   });
@@ -200,11 +84,8 @@ describe("# Articles endpoints", () => {
       .set("Authorization", `Bearer ${token}`)
       .send(rest)
       .end((err, res) => {
-        expect(res.status).equals(OK);
-        expect(res.body.message).to.contain(
-          "Article ready to be posted on facebook"
-        );
-        expect(res.body.response.spawnfile).to.equal("open");
+        expect(res.status).equals(BAD_REQUEST);
+        expect(res.body.message).to.contain('"body" is required');
         done();
       });
   });
@@ -217,11 +98,8 @@ describe("# Articles endpoints", () => {
       .set("Authorization", `Bearer ${token}`)
       .send(rest)
       .end((err, res) => {
-        expect(res.status).equals(OK);
-        expect(res.body.message).to.contain(
-          "Article ready to be posted on linkedIn"
-        );
-        expect(res.body.response.spawnfile).to.equal("open");
+        expect(res.status).equals(BAD_REQUEST);
+        expect(res.body.message).to.contain('"description" is required');
         done();
       });
   });
@@ -234,9 +112,12 @@ describe("# Articles endpoints", () => {
       .set("Authorization", `Bearer ${token}`)
       .send(rest)
       .end((err, res) => {
-        expect(res.status).equals(OK);
-        expect(res.body.message).to.contain("Article ready to be posted on Email");
-        expect(res.body.response.spawnfile).to.equal("open");
+        expect(res.status).equals(CREATED);
+        expect(res.body.message).to.contain("Article created");
+        expect(res.body).to.haveOwnProperty("article");
+        expect(res.body.article).to.haveOwnProperty("createdAt");
+        expect(res.body.article).to.haveOwnProperty("title");
+        expect(res.body.article.title).to.contain(newArticle.title);
         done();
       });
   });
